+++ malleefowl/processes/wps_swift.py
@@ -1,5 +1,5 @@
 from malleefowl.process import WPSProcess
-from malleefowl import cloud
+from malleefowl import swiftcloud
 
 from malleefowl import config
 
@@ -33,10 +33,10 @@ class SwiftProcess(WPSProcess):
             )
 
 
-class Login(WPSProcess):
+class SwiftLogin(WPSProcess):
     def __init__(self):
         WPSProcess.__init__(self,
-            identifier = "cloud_login",
+            identifier = "swift_login",
             title = "Login to Swift Cloud",
             version = "1.0",
             abstract="Login to Swift Cloud and get Token.")
@@ -89,7 +89,7 @@ class Login(WPSProcess):
             )
 
     def execute(self):
-        (storage_url, auth_token) = cloud.login(
+        (storage_url, auth_token) = swiftcloud.login(
             self.username.getValue(),
             self.password.getValue(),
             auth_url = self.auth_url.getValue(),
@@ -98,13 +98,13 @@ class Login(WPSProcess):
         self.storage_url.setValue( storage_url )
         self.auth_token.setValue( auth_token )
 
-class Download(SwiftProcess):
+class SwiftDownload(SwiftProcess):
     def __init__(self):
         SwiftProcess.__init__(self,
-            identifier = "cloud_download",
+            identifier = "swift_download",
             title = "Download files from Swift Cloud",
             version = "1.0",
-            abstract="Downloads files from Swift Cloud and provides file list as json document.")       
+            abstract="Downloads files from Swift Cloud and provides file List as JSON Document.")       
 
         self.container = self.addLiteralInput(
             identifier = "container",
@@ -134,7 +134,7 @@ class Download(SwiftProcess):
             )
 
     def execute(self):
-        files = cloud.download(
+        files = swiftcloud.download(
             self.storage_url.getValue(),
             self.auth_token.getValue(),
             self.container.getValue(),
@@ -147,10 +147,10 @@ class Download(SwiftProcess):
         self.output.setValue( outfile )
 
 
-class DownloadUrls(SwiftProcess):
+class SwiftDownloadUrls(SwiftProcess):
     def __init__(self):
         SwiftProcess.__init__(self,
-            identifier = "cloud_download_urls",
+            identifier = "swift_download_urls",
             title = "Provide download URLs for files from Swift Cloud",
             version = "1.0",
             abstract="Provide download URLs for files from Swift Cloud and return url list as json document.")
@@ -183,7 +183,7 @@ class DownloadUrls(SwiftProcess):
             )
 
     def execute(self):
-        files = cloud.download_urls(
+        files = swiftcloud.download_urls(
             self.storage_url.getValue(),
             self.auth_token.getValue(),
             self.container.getValue(),